src/backend/requirements.txt
fastapi
uvicorn
python-socketio